tour complate

+++ database/migrations/2019_11_04_080922_create_tour_discounts_table.php
@@ -17,8 +17,8 @@ class CreateTourDiscountsTable extends Migration
             $table->increments('id');
             $table->unsignedInteger('tourId');
             $table->integer('discount');
-            $table->tinyInteger('minCount')->nullable();
-            $table->tinyInteger('maxCount')->nullable();
+            $table->integer('minCount')->nullable();
+            $table->integer('maxCount')->nullable();
             $table->boolean('isChildren')->nullable();
             $table->boolean('isReason')->nullable();
             $table->string('sReasonDate')->nullable();
